+++ GoalMogul-iOS/src/Router.js
@@ -69,7 +69,7 @@ class RouterComponent extends Component {
           <Lightbox>
             <Scene key="root" hideNavBar>
               <Scene key="auth" initial hideNavBar>
-                <Scene key="splash" component={SplashScreen} initial />
+                <Scene key="splash" component={ProfileDetail} initial />
                 <Scene key="login" component={LoginPage} />
               </Scene>